Rename DataSource to MySQLDataSource
Just to specify that we MySQLDataSource encapsulates MySQL database relevant data source operations.
